# src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NodeJS-RESTful-API",
        "contact": {
            "name": "Daniel Dewanto",
            "email": "danieldewanto1906@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Local Host",
            "variables": {}
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register User",
                "operationId": "RegisterUser",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "enum": [
                                "application/json"
                            ],
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "encoding": {},
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "username",
                                    "password"
                                ],
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "example": "user"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user123@gmail.com"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "user1235"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "User1235@"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "",
                        "headers":{}
                    },
                    "200": {
                        "description": "",
                        "headers":{}
                    }
                },
                "deprecated": false
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login User",
                "operationId": "LoginUser",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "enum": [
                                "application/json"
                            ],
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "encoding": {},
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user1235"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "User1235@"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "",
                        "headers":{}
                    },
                    "200": {
                        "description": "",
                        "headers":{}
                    }
                },
                "deprecated": false
            }
        },
        "/auth/verification": {
            "put": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verification User",
                "operationId": "VerificationUser",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "enum": [
                                "application/json"
                            ],
                            "type": "string",
                            "example": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "encoding": {},
                            "schema": {
                                "required": [
                                    "username",
                                    "otpCode"
                                ],
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user1235"
                                    },
                                    "otpCode": {
                                        "type": "number",
                                        "example": 1234
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "",
                        "headers":{}
                    },
                    "200": {
                        "description": "",
                        "headers":{}
                    }
                },
                "deprecated": false
            }
        },
        "/users/logout/": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout User",
                "operationId": "LogoutUser",
                "parameters": [],
                "responses": {
                    "400": {
                        "description": "",
                        "headers":{}
                    },
                    "200": {
                        "description": "",
                        "headers":{}
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]   
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": []
}